game page initial implementation
