+++ src/App.js
@@ -1,7 +1,13 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
-import { SplashPage, CreateGame, JoinGame, Footer } from "./components";
+import {
+  SplashPage,
+  CreateGame,
+  JoinGame,
+  GamePage,
+} from "./components/routes";
+import { Footer } from "./components/library";
 
 import "./App.css";
 
@@ -12,6 +18,7 @@ const App = () => (
         <Route path="/" element={<SplashPage />} />
         <Route path="create-game" element={<CreateGame />} />
         <Route path="join-game" element={<JoinGame />} />
+        <Route path="game-page" element={<GamePage />} />
       </Routes>
       <Footer />
     </Router>
