--- src/components/footer/Footer.js
@@ -1,17 +1,0 @@
-import React from "react";
-import Typography from "@mui/material/Typography";
-
-import "./Footer.css";
-
-const Footer = () => (
-  <div>
-    <div className="footer-container">
-      <Typography align="center" variant="p" className="footer-subtitle">
-        Ioannis Siampalias
-      </Typography>
-    </div>
-  </div>
-);
-
-export { Footer };
-export default Footer;
