--- src/components/index.js
@@ -1,8 +1,0 @@
-import CtaButton from "./cta-button";
-import Image from "./image";
-import SplashPage from "./splash-page";
-import CreateGame from "./create-game";
-import JoinGame from "./join-game";
-import Footer from "./footer";
-
-export { CtaButton, Image, SplashPage, CreateGame, JoinGame, Footer };
