+++ src/components/library/footer/Footer.css
@@ -7,7 +7,6 @@
   position: absolute;
   bottom: 0;
 }
-
 .footer-subtitle {
   color: white;
 }
