+++ src/components/library/image/index.js
