+++ src/components/library/index.js
@@ -1,0 +1,5 @@
+import CtaButton from "./cta-button";
+import Image from "./image";
+import Footer from "./footer";
+
+export { CtaButton, Image, Footer };
