+++ src/components/routes/create-game/CreateGame.css
@@ -1,21 +1,16 @@
 .items-container {
-  display: flex;
-  justify-content: space-between;
   padding: 32px;
   border: 10px solid #9c27b0;
   border-radius: 5px;
   background-color: rgba(0, 0, 0, 0.5);
 }
-
-#join-textfield {
+#create-textfield {
   border-radius: 5px;
-  background-color: #ed6c02;
+  background-color: #9c27b0;
 }
-
 .MuiTextField-root > label {
   color: white;
 }
-
 .back-button {
   position: absolute !important;
   right: 0;
