+++ src/components/routes/game-page/index.js
@@ -1,0 +1,3 @@
+import GamePage from "./GamePage";
+
+export default GamePage;
