+++ src/components/routes/index.js
@@ -1,0 +1,6 @@
+import SplashPage from "./splash-page";
+import CreateGame from "./create-game";
+import JoinGame from "./join-game";
+import GamePage from "./game-page";
+
+export { SplashPage, CreateGame, JoinGame, GamePage };
