+++ src/components/routes/splash-page/SplashPage.css
@@ -12,12 +12,6 @@
   color: white;
   padding-bottom: 32px;
 }
-.button-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
 .button {
   width: 320px;
   height: 80px;
