+++ src/components/routes/splash-page/index.js
